Fix for issue #14: Weaken type checks when retrieving datastore entities.
When reconstructing datastore entities, a warning is written to the logger
instead of raising a NoSuchPropertyError

Also added a top level 'datastoreLogger' value so that direct access to a datastore instance is not required in order to log records. Datastore.logger redirects to this, and the logger name is now configurable.